+++ app/src/main/java/com/mensinator/app/ui/navigation/BarItem.kt
@@ -1,4 +1,4 @@
-package com.mensinator.app.navigation
+package com.mensinator.app.ui.navigation
 
 data class BarItem(
     val screen: Screen,
